// frontend/src/backend.d.ts
export type Time = bigint;
// ...
export interface CryptoAsset {
    name: string;
    usdValue: number;
    symbol: string;
}
// ...
export interface CapitalFlow {
    marketImpact: number;
    flowIntensity: number;
    pnlRatio: number;
    timestamp: Time;
    amount: number;
    fromAsset: CryptoAsset;
    toAsset: CryptoAsset;
}